requirements.txt
pylint
pytest
python-dotenv
boto3
psycopg2
psycopg2-binary